Add RPC

+++ server.js
@@ -4,16 +4,21 @@ if(cluster.isMaster) {
   const fs = require('fs')
   const procNumber = require('os').cpus().length
   const net = require('net');
+  const rpc = require('rpc-stream');
 
-  // basic echo server. It also appends a signature string.
+  // basic echo server executing a Remote Procedure Call.
+  // It also appends a signature string.
   const server = net.createServer(s => {
-    s.on('data', data => {
-      s.write( data.toString() + ' <-- [from server]')
-    })
+    const server = rpc({addTail: function(data, cb){
+      cb(null, data + ' <--[from rpc]');
+    }})
+    server.pipe(s).pipe(server);
   })
 
-  // remove the old socket files or they will block the app.
-  fs.unlinkSync(__dirname + '/tmp.sock');
+  // remove the old socket file or it will block the app.
+  if(fs.existsSync(__dirname + '/tmp.sock')) {
+    fs.unlinkSync(__dirname + '/tmp.sock');
+  }
 
   // the server in the master process is listening to a UNIX socket
   server.listen(__dirname + '/tmp.sock', () => {
@@ -22,12 +27,12 @@ if(cluster.isMaster) {
       cluster.fork();
   });
 } else {
-  // start an worker process running an express app
+  // start a worker process running an express app
   startWorker(process.pid);
 }
 
 
-// helper functions
+// worker function
 function startWorker(id) {
 
   console.log('starting worker', id);
@@ -35,35 +40,29 @@ function startWorker(id) {
   var app = express();
   const net = require('net');
 
+  // connect the child process to execute the RPC
+  const rpc = require('rpc-stream');
+  client = rpc();
+  const remote = client.wrap('addTail');
+  const conn = net.connect({path: __dirname + '/tmp.sock'})
+  client.pipe(conn).pipe(client);
+
+
   app.get("/", function (req, res) {
-    res.type('txt').send('cluster and interprocess communication test \n\
-hit the "\\test" endpoint')
+    res.type('txt').send('cluster and interprocess rpc test \n\
+hit the "/test" endpoint')
   });
 
   app.get('/test', function(req, res){
-    writeToSocket(net, id, (data) => {
+    // execute the rpc, and send back its response
+    remote.addTail(`hello from pid ${id}`, (err, data) => {
+      if(err) return res.status(500).send('error');
       res.type('txt').send(data);
     });
   })
 
   // listen for requests :)
-  var listener = app.listen(process.env.PORT, function () {
+  var listener = app.listen(process.env.PORT || 3000, function () {
     console.log('Your app is listening on port ' + listener.address().port);
-  });
-}
-
-function writeToSocket(net, pid, cb) {
-
-    const client = net.connect({path: __dirname + '/tmp.sock'}, () => {
-    //'connect' listener
-    console.log(`${pid} connected`);
-    client.write(`hello from pid ${pid}`);
-  });
-  client.on('data', (data) => {
-    cb(data.toString());
-    client.end();
-  });
-  client.on('end', () => {
-    console.log(`${pid} disconnected`);
   });
 }
